LayoutTests/fast/filesystem/directory-entry-to-uri-expected.txt

Obtaining URI from DirectoryEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDirectoryURI is 'filesystem:file:///temporary/testDirectory'
PASS successfullyParsed is true

TEST COMPLETE
